# docs-site/contracts/product-http.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WuKongIM HTTP API",
    "version": "3.0.0-beta-snapshot",
    "description": "The 48 core WuKongIM HTTP API JSON operations, plus the plugin-defined ANY route in x-wukongim-extension-routes. Routes have no built-in authentication.",
    "x-i18n": {
      "zh": {
        "title": "WuKongIM HTTP API",
        "description": "与当前运行时的 48 个核心 WuKongIM HTTP API JSON 操作对齐；插件定义的 ANY 路由见 x-wukongim-extension-routes。这些路由没有内建鉴权。"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5001",
      "description": "Local single-node cluster example",
      "x-i18n": {
        "zh": {
          "description": "本地单节点集群示例"
        }
      }
    }
  ],
...
  "security": [],
  "paths": {
...
    "/channel": {
      "post": {
        "operationId": "upsertChannel",
        "summary": "Create or update Channel metadata",
        "description": "Upserts all metadata flags and optionally resets subscribers; disband is terminal.",
        "x-i18n": {
          "zh": {
            "summary": "创建或更新 Channel 元数据",
            "description": "全量更新元数据标志并可重置订阅者；disband 为终态。"
          }
        },
        "tags": [
          "Channels"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelUpsert"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/info": {
      "post": {
        "operationId": "updateChannelInfo",
        "summary": "Replace Channel metadata",
        "description": "Passes the zero-valued full record to UpdateInfo when fields are omitted; retained for compatibility.",
        "x-i18n": {
          "zh": {
            "summary": "替换 Channel 元数据",
            "description": "字段省略时仍把零值完整记录交给 UpdateInfo；该接口仅为兼容保留。"
          }
        },
        "tags": [
          "Channels"
        ],
        "deprecated": true,
        "security": [],
        "x-wukongim-trust": "operator-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelInfo"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/delete": {
      "post": {
        "operationId": "disbandChannel",
        "summary": "Terminally disband a Channel",
        "description": "Sets the durable disband flag without deleting Channel identity; request-key validation is weak.",
        "x-i18n": {
          "zh": {
            "summary": "终态解散 Channel",
            "description": "设置持久化 disband 标志但不删除 Channel 身份；请求 Key 校验较弱。"
          }
        },
        "tags": [
          "Channels"
        ],
        "deprecated": true,
        "security": [],
        "x-wukongim-trust": "operator-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/WeakChannelKey"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/subscriber_add": {
      "post": {
        "operationId": "addChannelSubscribers",
        "summary": "Add or replace durable subscribers",
        "description": "Adds non-blank subscribers; channel_type 0 becomes group type 2 and reset=1 replaces the snapshot.",
        "x-i18n": {
          "zh": {
            "summary": "添加或替换持久订阅者",
            "description": "添加非空白订阅者；channel_type=0 转为群组类型 2，reset=1 替换快照。"
          }
        },
        "tags": [
          "Channels"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelSubscriberAdd"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/subscriber_remove": {
      "post": {
        "operationId": "removeChannelSubscribers",
        "summary": "Remove durable subscribers",
        "description": "Removes non-blank subscribers; unlike subscriber_add, channel_type 0 is not normalized.",
        "x-i18n": {
          "zh": {
            "summary": "移除持久订阅者",
            "description": "移除非空白订阅者；与 subscriber_add 不同，channel_type=0 不会被归一化。"
          }
        },
        "tags": [
          "Channels"
        ],
        "deprecated": true,
        "security": [],
        "x-wukongim-trust": "operator-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelSubscriberRemove"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/subscriber_remove_all": {
      "post": {
        "operationId": "removeAllChannelSubscribers",
        "summary": "Remove all durable subscribers",
        "description": "Clears ordinary subscribers through bounded internal pages; person Channels are rejected.",
        "x-i18n": {
          "zh": {
            "summary": "移除全部持久订阅者",
            "description": "通过内部有界分页清空普通订阅者；个人 Channel 会被拒绝。"
          }
        },
        "tags": [
          "Channels"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/NonPersonChannelKey"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/tmpchannel/subscriber_set": {
      "post": {
        "operationId": "setTemporaryChannelSubscribers",
        "summary": "Replace temporary Channel subscribers",
        "description": "Replaces the derived Type-8 temporary subscriber list; UID elements are not validated by the entry adapter.",
        "x-i18n": {
          "zh": {
            "summary": "替换临时 Channel 订阅者",
            "description": "替换派生 Type-8 临时订阅者列表；入口适配器不校验 UID 元素。"
          }
        },
        "tags": [
          "Channels"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/TemporarySubscribers"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/blacklist_add": {
      "post": {
        "operationId": "addChannelDenylistMembers",
        "summary": "Add Channel denylist members",
        "description": "Adds the supplied UID strings to the derived Channel denylist.",
        "x-i18n": {
          "zh": {
            "summary": "添加 Channel 拒绝列表成员",
            "description": "把给定 UID 字符串加入派生 Channel 拒绝列表。"
          }
        },
        "tags": [
          "Channels"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelMembers"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/blacklist_set": {
      "post": {
        "operationId": "setChannelDenylistMembers",
        "summary": "Replace Channel denylist members",
        "description": "Removes the old list before adding the supplied values; only channel_id is validated at entry.",
        "x-i18n": {
          "zh": {
            "summary": "替换 Channel 拒绝列表",
            "description": "先移除旧列表再添加给定值；入口只校验 channel_id。"
          }
        },
        "tags": [
          "Channels"
        ],
        "deprecated": true,
        "security": [],
        "x-wukongim-trust": "operator-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelMemberSet"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/blacklist_remove": {
      "post": {
        "operationId": "removeChannelDenylistMembers",
        "summary": "Remove Channel denylist members",
        "description": "Removes the supplied UID strings from the derived Channel denylist.",
        "x-i18n": {
          "zh": {
            "summary": "移除 Channel 拒绝列表成员",
            "description": "从派生 Channel 拒绝列表移除给定 UID 字符串。"
          }
        },
        "tags": [
          "Channels"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelMembers"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/blacklist_remove_all": {
      "post": {
        "operationId": "removeAllChannelDenylistMembers",
        "summary": "Remove all Channel denylist members",
        "description": "Clears the derived denylist through bounded internal traversal.",
        "x-i18n": {
          "zh": {
            "summary": "移除全部 Channel 拒绝列表成员",
            "description": "通过内部有界遍历清空派生拒绝列表。"
          }
        },
        "tags": [
          "Channels"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelKey"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/whitelist_add": {
      "post": {
        "operationId": "addChannelAllowlistMembers",
        "summary": "Add Channel allowlist members",
        "description": "Adds non-blank UIDs to the derived Channel allowlist.",
        "x-i18n": {
          "zh": {
            "summary": "添加 Channel 允许列表成员",
            "description": "把非空白 UID 加入派生 Channel 允许列表。"
          }
        },
        "tags": [
          "Channels"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelAllowlistMembers"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/whitelist_set": {
      "post": {
        "operationId": "setChannelAllowlistMembers",
        "summary": "Replace Channel allowlist members",
        "description": "Removes the old list before adding supplied values; only channel_id is validated at entry.",
        "x-i18n": {
          "zh": {
            "summary": "替换 Channel 允许列表",
            "description": "先移除旧列表再添加给定值；入口只校验 channel_id。"
          }
        },
        "tags": [
          "Channels"
        ],
        "deprecated": true,
        "security": [],
        "x-wukongim-trust": "operator-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelMemberSet"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/whitelist_remove": {
      "post": {
        "operationId": "removeChannelAllowlistMembers",
        "summary": "Remove Channel allowlist members",
        "description": "Removes non-blank UIDs from the derived Channel allowlist.",
        "x-i18n": {
          "zh": {
            "summary": "移除 Channel 允许列表成员",
            "description": "从派生 Channel 允许列表移除非空白 UID。"
          }
        },
        "tags": [
          "Channels"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelAllowlistMembers"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/whitelist_remove_all": {
      "post": {
        "operationId": "removeAllChannelAllowlistMembers",
        "summary": "Remove all Channel allowlist members",
        "description": "Clears the derived allowlist through bounded internal traversal.",
        "x-i18n": {
          "zh": {
            "summary": "移除全部 Channel 允许列表成员",
            "description": "通过内部有界遍历清空派生允许列表。"
          }
        },
        "tags": [
          "Channels"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelKey"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
    "/channel/whitelist": {
      "get": {
        "operationId": "listChannelAllowlistMembers",
        "summary": "List all Channel allowlist members",
        "description": "Returns an unbounded full list; missing or invalid channel_type is silently treated as 0.",
        "x-i18n": {
          "zh": {
            "summary": "列出全部 Channel 允许列表成员",
            "description": "返回无界完整列表；channel_type 缺失或无效时会静默视为 0。"
          }
        },
        "tags": [
          "Channels"
        ],
        "deprecated": true,
        "security": [],
        "x-wukongim-trust": "operator-only",
        "parameters": [
          {
            "$ref": "#/components/parameters/ChannelIDQuery"
          },
          {
            "$ref": "#/components/parameters/ChannelTypeQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ChannelMemberListOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    },
...
    "/channel/expire": {
      "post": {
        "operationId": "updateChannelExpiry",
        "summary": "Update channel expiry",
        "description": "Persist expiry through Channel authority; personal channels may be created.",
        "x-i18n": {
          "zh": {
            "summary": "更新频道过期时间",
            "description": "通过频道权威持久化过期时间；个人频道可按需创建。"
          }
        },
        "tags": [
          "Channels"
        ],
        "security": [],
        "x-wukongim-trust": "trusted-backend-only",
        "requestBody": {
          "$ref": "#/components/requestBodies/ChannelExpiry"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/StatusOK"
          },
          "400": {
            "$ref": "#/components/responses/CompatibilityBadRequest"
          },
          "503": {
            "$ref": "#/components/responses/MaintenanceUnavailable"
          }
        }
      }
    }
  },
  "components": {
...
    "requestBodies": {
...
      "ChannelInfo": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ChannelInfoRequest"
            }
          }
        }
      },
      "ChannelUpsert": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ChannelUpsertRequest"
            }
          }
        }
      },
      "WeakChannelKey": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/WeakChannelKeyRequest"
            }
          }
        }
      },
      "ChannelSubscriberAdd": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ChannelSubscriberAddRequest"
            }
          }
        }
      },
      "ChannelSubscriberRemove": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ChannelSubscriberRemoveRequest"
            }
          }
        }
      },
      "NonPersonChannelKey": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NonPersonChannelKeyRequest"
            }
          }
        }
      },
      "TemporarySubscribers": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TemporarySubscriberSetRequest"
            }
          }
        }
      },
...
    },
    "responses": {
      "StatusOK": {
        "description": "Mutation accepted",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StatusEnvelope"
            }
          }
        }
      },
      "CompatibilityBadRequest": {
        "description": "Input, validation, dependency, or downstream error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CompatibilityError"
            }
          }
        }
      },
      "MaintenanceUnavailable": {
        "description": "Controller restore maintenance is active",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MaintenanceError"
            }
          }
        }
      },
...
    },
...
  }
}